extract unix commands into its own package
